dont remember

+++ rewrite/datasets/Webcams.py
@@ -47,14 +47,14 @@ class Webcams_reg(Dataset):
         dims = (data.size(1)-50, data.size(2)-6)
         data = f.crop(data, 46, 2, dims[0], dims[1])
         data = self.transform(data)
-        data = data.float()
+        data = data.to(torch.float32)
         
         string_value = path.basename(img_path)
         string_value = string_value.split('_')[2].split('.')[0].split('S')[1].split('m')[0].replace('-', '.')
         float_value = float(string_value)
         float_value = 10.0 if float_value >= 10.0 else float_value
         
-        value = torch.Tensor([float_value])
+        value = torch.Tensor([float_value]).to(torch.float32)
         
         return (data, value)
